Check if email and username is available for member and society registration
