+++ src/app/api/register/society/route.ts
@@ -1,18 +1,63 @@
 import { NextRequest, NextResponse } from "next/server";
-import { User } from "../../../../../db/schema";
+import { Society } from "../../../../../db/schema";
 import argon2 from "argon2";
 
-export async function POST(req: NextRequest) {
-  const { name, username, email, password, mobile } = (await req.json()) as any;
+export async function GET(req: NextRequest) {
+  try {
+    const { searchParams } = new URL(req.url);
+    const username = searchParams.get("username");
+    const email = searchParams.get("email");
 
-  const society = await User.create({
-    name,
-    username,
-    mobile,
-    email,
-    password: await argon2.hash(password),
-  });
-  console.log(society);
+    if (email) {
+      const userEmailExists = await Society.findOne({ email });
+      const societyEmailExists = await Society.findOne({ email });
+
+      return NextResponse.json({
+        emailExists: !!(userEmailExists || societyEmailExists),
+      });
+    }
+
+    if (username) {
+      const userUsernameExists = await Society.findOne({ username });
+      const societyUsernameExists = await Society.findOne({ username });
+
+      return NextResponse.json({
+        usernameExists: !!(userUsernameExists || societyUsernameExists),
+      });
+    }
+
+    return NextResponse.json(
+      { error: "Please provide 'username' or 'email' to check." },
+      { status: 400 }
+    );
+  } catch (e) {
+    console.error("Validation error:", e);
+    return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
+  }
+}
+
+export async function POST(req: NextRequest) {
+  const { name, username, email, password } = (await req.json()) as any;
+  try {
+    if (!name || !username || !email || !password) {
+      console.error("Missing entries");
+      return NextResponse.json("Invalid Entry");
+    }
+    const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
+    if (!emailRegex.test(email)) {
+      return NextResponse.json("Invalid Email Format!");
+    }
+
+    const society = await Society.create({
+      name,
+      username,
+      email,
+      password: await argon2.hash(password),
+    });
+    console.log(society);
+  } catch (e) {
+    NextResponse.json("Error!");
+  }
 
   return NextResponse.json({ ok: true });
 }
